changes done

+++ com/eurekaAccounts/stocks/DAO/LookupStockFundamentsDAO.java
@@ -32,7 +32,7 @@ public class LookupStockFundamentsDAO extends BaseDAO {
         //System.out.println(resultSet);
         while (resultSet.next()) {
             StocksectorVO stockFundementalsVO = new StocksectorVO();
-            stockFundementalsVO.setSectorId(resultSet.getInt("sector_id"));
+            stockFundementalsVO.setSectorId(resultSet.getBigDecimal("sector_id"));
             stockFundementalsVO.setSubSectorId(resultSet.getInt("subsector_id"));
             stockFundementalsVO.setTickerSymbol(resultSet.getString("ticker_symbol"));
             stockFundementalsVO.setMarketCap(resultSet.getLong("market_cap"));
